+++ packages/partials/concurrent/src/index.ts
@@ -15,6 +15,7 @@
  */
 
 export * from './Classes/ThrottleController';
+export * from './Classes/BatchBuffer';
 export * from './Classes/DebounceController';
 export * from './Classes/SlideWindowCounter';
 export * from './Classes/CountingRateLimiter';